Restructure devtools code
